+++ hydra-cardano-api/hydra-cardano-api.cabal
@@ -90,6 +90,7 @@ library
     Hydra.Cardano.Api.PlutusScriptVersion
     Hydra.Cardano.Api.PolicyId
     Hydra.Cardano.Api.Prelude
+    Hydra.Cardano.Api.Pretty
     Hydra.Cardano.Api.ReferenceScript
     Hydra.Cardano.Api.ReferenceTxInsScriptsInlineDatumsSupportedInEra
     Hydra.Cardano.Api.ScriptData
